Refactoring DatabaseConfiguration : Moving it to the core project for better structure
